Add DependenceTranslationFilter
Move DependenceTranslationFilter from CensusDataFilter to specific
filter.
